pyproject.toml
[tool.poetry]
name = "kolesa_parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
click = "^8.1.3"
Scrapy = "^2.6.3"
scrapy-fake-useragent = "^1.4.4"
scrapy-rotating-proxies = "^0.6.2"
scrapy-user-agents = "^0.1.1"
google-re2 = "^0.2.20220601"
SQLAlchemy = "^1.4.41"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
